+++ CHANGES.rst
@@ -19,7 +19,7 @@ Features added
 * #13439: linkcheck: Permit warning on every redirect with
   ``linkcheck_allowed_redirects = {}``.
   Patch by Adam Turner.
-* Extend duration extension with new configuration options.
+* #13468: Add config options to duration extension.
 
 Bugs fixed
 ----------